Not use float32 on API

+++ api/canvas/canvas.go
@@ -4,25 +4,25 @@ package canvas
 func Begin()
 
 //go:wasmimport sunani canvas.clear
-func Clear(r, g, b, a float32)
+func Clear(r, g, b, a uint32)
 
 //go:wasmimport sunani canvas.color
-func Color(r, g, b, a float32)
+func Color(r, g, b, a uint32)
 
 //go:wasmimport sunani canvas.line
-func Line(x1, y1 float32, x2, y2 float32)
+func Line(x1, y1 uint32, x2, y2 uint32)
 
 //go:wasmimport sunani canvas.rect
-func Rect(x, y float32, w, h float32)
+func Rect(x, y uint32, w, h uint32)
 
 //go:wasmimport sunani canvas.fill_rect
-func FillRect(x, y float32, w, h float32)
+func FillRect(x, y uint32, w, h uint32)
 
 //go:wasmimport sunani canvas.path
-func Path(x, y float32)
+func Path(x, y uint32)
 
 //go:wasmimport sunani canvas.vertex
-func Vertex(x, y float32)
+func Vertex(x, y uint32)
 
 //go:wasmimport sunani canvas.polygon
 func Polygon()
